fixed unique name form values
